requirements.txt
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==41.0.1
django-environ==0.10.0
et-xmlfile==1.1.0
idna==3.4
msal==1.22.0
Office365-REST-Python-Client==2.4.1
openpyxl==3.0.10
pycparser==2.21
PyJWT==2.7.0
pytz==2023.3
requests==2.31.0
urllib3==2.0.3